integer overflow
# message auto-generated for no-merge-commit merge:
merge integer_overflow into br_develop_boostio

totalLen累加数据源为对端返回数据，没有做任何校验，存在整数溢出风险


Match-id-671f2554ec133b2ebbdd110cf2b08645bb4112fa

+++ src/sdk/mirror_client.cpp
@@ -1036,7 +1036,7 @@ BResult MirrorClient::DataCopy(const char *from, uint32_t fromLen, SliceAddrDesc
     uint32_t addrNum)
 {
     // 检查slice内存大小是否大于等于fromLen.
-    uint32_t totalLen = 0;
+    uint64_t totalLen = 0;
     for (uint32_t i = 0; i < addrNum; i++) {
         totalLen += addr[i].chunkLen;
     }
